Creating member.js
Lists the tag/username/id/nickname(type) of members with given role, limited to maxlength in embed.
tmp\memberlist.txt is a file that this command uses to store data. Whenever the command is used again, the txt file is overwritten.

Default type is set to "nickname".
Default maxlength is set to 10.